requirements.txt
rumps
psutil
requests 
mss
Pillow
openai
certifi
httpx
pyobjc-core
pyobjc-framework-Quartz 